moved logs to inside check
