+++ otter/check/__init__.py
@@ -7,9 +7,9 @@ import os
 from glob import glob
 from jinja2 import Template
 
+from .logs import LogEntry, EventType
 from .notebook import _OTTER_LOG_FILENAME
 from ..execute import grade_notebook, check
-from ..logs import LogEntry, EventType
 from ..utils import block_print
